disentangle metrics module: remove group_level.py and move pooled feature metrics inside feature_level instead; move dedicated feature & observation level qc metrics out of feature_metrics into qc.py
