+++ src/alphapepttools/metrics/__init__.py
@@ -1,14 +1,17 @@
 """Metrics for the quality assessment of the analysis"""
 
 from .feature_level import (
-    calculate_qc_metrics,
     coefficient_of_variation,
+    pooled_coefficient_of_variation,
+    pooled_median_absolute_deviation,
+)
+from .principal_component_regression import principal_component_regression
+from .qc import (
+    calculate_qc_metrics,
     fraction_complete,
     number_detected,
     total_intensity,
 )
-from .group_level import pooled_coefficient_of_variation, pooled_median_absolute_deviation
-from .principal_component_regression import principal_component_regression
 
 __all__ = [
     "calculate_qc_metrics",
